Unbreak rubocop for regular usage
This project already has `.rubocop.yml` and `.rubocop_todo.yml`, but it
doesn't track what version it cares about and those files are no longer
valid for a contemporary version.  Added to `Gemfile` and lock file.

38 different cops have changed names since the last time we checked, apparently.  Thankfully, only one of
them is in `.rubocop.yml`, so after renaming that and an `--auto-gen-config` we get:

```
86 files inspected, 4370 offenses detected
```

Sounds about right for code this old.