Treat should_register as always false
Note the abundance of follow-on entailments, i.e. large swaths of code
that only ran or auxilliary configs that were only honored when
`should_register`'s value was in question.

This simplification is enabled by having removed `bin` and `devel`
scripts and therefore only needing to support the limited configs
documented for the web UI.