remove validate_bundle_dir config option, along with supporting code and gem inclusion
twas something of a security risk before, would be an even worse security risk inside a webapp, and it's not used much or at all anymore anyway (per @blalbrit).  if/when we reunify the command-line functionality from the v3-legacy branch into master, we can reimplement something like this in a safer/saner way (e.g. scripts get checked into the pre-assembly codebase, and are added to an enumerated list of validation scripts that can be run).