BundleContext initialization has some requirements
This is not exhaustive, but it applies some defaults and some strict
requirements up front.  After the AR conversion, these will change to AR
validation components, but for now the takeaways are:
- `bundle_dir` is required
- a bunch of empty hash defaults at the top level
- `content_md_creation.style` defaults to `:default`, logically
- `project_style.content_structure` defaults to `:simple_image`, per Ben
- use built in `Hash#transform_values!` instead of opaque `Assembly::Utils` invocation

There are bound to be other validation requirements, but this is a step
in the right direction for now.

Cherry-picked over (conflicts resolved).