Display Globus errors
Adjust the globus controller to check how many globus destinations a
user has already prior to creating another one. If they have 99 or more
the controller returns an JSON error message with 400 status code.

The Stimulus controller checks the response and replaces the Create
Globus Link button with the error message if it's a 400. If it's another
type of error (potentially a 500 error with no JSON body) it uses the
generic message 'Unexpected Globus Error'.

closes #1340