Use globus endpoint id to extract path from URL
Depending on how the user arrives at Globus (either from Preassembly, or
from the Globus email, or navigating directly themselves) the Globus
path embedded in the URL could be in either the destination_path or
origin_path paremeters. This commit adds the ability to use the Globus
Destination ID configured in app settings to infer where to look for the
Globus path.

Fixes #1333